vlandev: fix type of vlan alias lookup result
The vlan alias kvlist stores unsigned int values (device_vlan_len,
config_init_vlan_entry), but vlandev_get_vid() read them through a
uint16_t pointer, which returns the two high-order bytes on big-endian
targets, i.e. vid 0 for every valid alias. Read the value with the
type it was stored as (vlan.c reads it through int *, which is
equivalent for valid vlan ids).

+++ vlandev.c
@@ -250,7 +250,7 @@ static uint16_t
 vlandev_get_vid(struct device *dev, const char *id_str)
 {
 	unsigned long id;
-	uint16_t *alias_id;
+	unsigned int *alias_id;
 	char *err;
 
 	id = strtoul(id_str, &err, 10);
